+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.6.3 (TBD)
  - [DICOM Importer] Display PTV ROI type in RT Structure tree.  Helpful for multi- PTV plans.
  - [DICOM Importer] Set force=True with pydicom.read_file() for XiO support  [issue #15](https://github.com/cutright/DVH-Analytics/issues/15)
  - [Query] Added a catch and error dialog for a MemoryError when querying if users queries more data than their system can handle
+ - [Database Connection] Initialize SQL tables upon successful new database connection
 
 v0.6.2 (2019.12.08)
 --------------------